Further adapt svyolr output to handle predict methods
svyolr output is very similar to polr output but missing two elements.
The model (data) and lp (linear predictor). Since the svyolr class was
coerced to polr to integrate into existing codebase, the model element is
required since it is used in the predict.Regression method. Commit adds
the model element to the svyolr list output.